refactor: angular folder structure
