+++ client/src/app/app.routes.ts
@@ -1,9 +1,9 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './home/home.component';
-import { LoginComponent } from './login/login.component';
-import { RegisterComponent } from './register/register.component';
+import { HomeComponent } from './components/home/home.component';
+import { LoginComponent } from './components/login/login.component';
+import { RegisterComponent } from './components/register/register.component';
 import { AuthGuard } from './auth.guard';
-import { NewSaleComponent } from './new-sale/new-sale.component';
+import { NewSaleComponent } from './components/new-sale/new-sale.component';
 
 export const routes: Routes = [
   { path: '', component: LoginComponent, title: 'Login' },
