+++ client/src/app/components/loader/loader.component.ts
