+++ client/src/app/components/new-sale/new-sale.component.html
