+++ client/src/app/components/register/register.component.ts
@@ -1,6 +1,6 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { env } from '../../../config/enviroments';
+import { env } from '../../../../config/enviroments';
 import axios from 'axios';
 import { CookieService } from 'ngx-cookie-service';
